+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Toolchain changes will now automatically reload the workspace ([#2196](https://github.com/swiftlang/vscode-swift/pull/2196))
+- Add **Clean Build Folder** automatically generated task to remove all build artifacts ([#2280](https://github.com/swiftlang/vscode-swift/pull/2280))
 
 ### Fixed