+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Toolchain changes will now automatically reload the workspace ([#2196](https://github.com/swiftlang/vscode-swift/pull/2196))
+- Add a button that installs Swiftly and the latest Swift toolchain to the extension's welcome walkthrough.
 
 ### Fixed
 
@@ -67,7 +68,6 @@
 ### Added
 
 - Add `swift.ignoreSwiftVersionFile` setting to ignore the toolchain set in `.swift-version` files ([#2073](https://github.com/swiftlang/vscode-swift/pull/2073))
-- Add a button that installs Swiftly and the latest Swift toolchain to the extension's welcome walkthrough.
 
 ### Fixed